New educational pages
These pages help users understand the changes related to the payment feature. Each page is intended for a different audience: users with a deprecated app and users with an updated app. Users will be engaged by using, respectively, a banner in the legacy wallet page and in the new payments page